pyproject.toml
[project]
name = "entropy"
version = "0.1.0"
description = "Personal random task managger"
readme = "README.md"
authors = [
    { name = "DeepeshKalura", email = "deepeshkalurs@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "click>=8.1.8",
    "pyfiglet>=1.0.2",
    "pygithub>=2.6.1",
    "rich-pixels>=3.0.1",
    "rich>=13.9.4",
    "sqlalchemy>=2.0.38",
]

[dependency-groups]
dev = [
    "pylint>=3.3.4",
    "pytest>=8.3.4",
    "python-dotenv>=1.0.1",
    "ruff>=0.9.7",
]
[project.scripts]
entropy = "src.act:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]


[tool.uv.workspace]
members = ["somn"]
